backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.31.1
sqlmodel==0.0.16
pydantic>=2.12.5
pydantic-settings>=2.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
better-exceptions==0.2.2
asyncpg==0.29.0
aiosqlite==0.19.0
greenlet>=3.0.0
alembic==1.13.1
python-multipart>=0.0.6
emails==0.6
celery==5.3.4
redis==5.0.1
python-dotenv==1.0.0

# Phase III Dependencies
# Dependency version constraints:
# - openai-agents 0.7.0 requires openai>=2.9.0
# - mcp 1.26.0 requires httpx>=0.27.1
openai-agents==0.7.0
openai>=2.9.0
mcp==1.26.0
httpx>=0.27.1
sse-starlette==2.2.1
dateparser==1.2.0
python-dateutil==2.8.2
pytest-asyncio==1.3.0